requirements.txt
Authlib==1.3.2
blinker==1.8.2
bokeh==3.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
dnspython==2.7.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tornado==6.4.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
xyzservices==2024.9.0
